Allow "KEYINFO --list"
Currently, this command is blocked.  GnuPG detects that the agent
connection is restricted and doesn't try to use it, while Sequoia
Chameleon does not implement the fallback and is unable to list secret
keys or decrypt messages.  Furthermore, gpg prints
"gpg: problem with fast path key listing: Forbidden - ignored", which
Mutt interprets as a prompt the user must respond to.  This causes the
user to need to press enter twice to send a signed email.

Fix these problems by allowing this request.  The request does not work
over a restricted connection, so an unrestricted connection must be
used.  However, the filtering done by split-gpg2 is far stronger than
the access checks in gpg-agent so there is no loss of security.

Fixes: QubesOS/qubes-issues#9529